added new files for week 4
Added hw-4, and practice sketch in my coursework folder. I began to experiment in the practice with ellipses. I wanted to make four circles in a column, but ended up with one strange skinny oval going down the page. I'm not sure why either. Is it overlapping? It's cool looking, so I'm going to keep it around. and try another.