Added comments and face shading
I'm struggling with this part. The shading is arc, but how do I make that line up with the edge of the ellipse from above? Scaling the ellipse didn't work, so I went back to the arc and managed to make it align by using the same size as the ellipse. Still doesn't look right though.