hw-4 experiment and image
I experimented with the chin shape, trying to add some shading in with long skinny rectangles. Still can't see the chin shape at all. I may have to rethink this how part with the chin.

Added a screenshot of my attempt. My  current thoughts? It looks stupid. Is this a face at all?