hw-10  random petal colors
I finally got the random petal colors to work when I moved the variable to the global to drawFlower function. Not sure its that proper place for it. I'll try to clean it up. ANyway, it works!

Sorta. I'd prefer to have each 'flower' on the grid to be a different color. Hmm that will be tricky.