update to reachedGoal
I fixed method for reachedGoal. I had to add in a radius to player.js and Goal. This radius controls their interaction.

I got stuck because I forgot to save my player.js file. No error, but when I checked the console.log for player.radius it came up as undefined. Then I saved the file with the radius defined and volia. The circles interact with each other now.