added avoidOthers funciton
I wanted the enemies to have to bounce off each other, or to avoid each other, but I was struggling with the math. So I just had them reappear randomly again. Then, instead of a game over, what if I try a scoring method. That everytime one touches you it adds up a score.